experiments/sockets/requirements.txt
absl-py==0.8.1
appdirs==1.4.3
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
cached-property==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
ConfigArgParse==1.2.1
configparser==4.0.2
cryptography==2.9
databases==0.2.6
docker==4.3.0.dev0
docker-compose==1.26.0.dev0
dockerpty==0.4.1
docopt==0.6.2
fastapi==0.54.1
Flask==1.1.2
gast==0.3.2
gevent==1.5a3
geventhttpclient-wheels==1.3.1.dev2
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.16.1
h11==0.9.0
h5py==2.9.0
httptools==0.0.13
idna==2.9
importlib-metadata==1.6.0
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
locust==0.0
locustio==0.14.5
loguru==0.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
minio==5.0.7
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
msgpack==1.0.0
mypy-extensions==0.4.3
numpy==1.18.1
opencv-contrib-python==4.2.0.32
paramiko==2.7.1
pathspec==0.7.0
Pillow==7.0.0
profilehooks==1.11.1
protobuf==3.11.2
psutil==5.7.0
psycopg2==2.8.4
pycparser==2.20
pydantic==1.3
pylint==2.4.4
pymongo==3.9.0
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.10.5
python-multipart==0.0.5
pytz==2019.3
PyYAML==5.2
pyzmq==19.0.0
redis==3.3.11
regex==2020.4.4
requests==2.23.0
rope==0.16.0
rq==1.2.0
scipy==1.3.2
six==1.13.0
SQLAlchemy==1.3.12
starlette==0.13.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
texttable==1.6.2
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
uvicorn==0.11.2
uvloop==0.14.0
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.1.0
